requirements.txt
fastapi
uvicorn
openai
httpx
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart